hwp-pico: move file to socs/agent/
